Update to btleplug
feature: change to btleplug backend

The previous library changed to an incompatible license,
so we are using a new library: btleplug.

It supports limiting scanning by service ID, which greatly
improved die acquisition speed, as well as supports a broader
range of OS's.

Unfortunately, since it's Rust, it doesn't produce binaries,
and I don't have the computers necessary to build macOS binaries,
so macOS is not supported until I can fix that.